+++ CHANGELOG.md
@@ -12,7 +12,7 @@
  *   Allow animate(), windowbuttonedit/create and windowimageedit/create to recieve functions
  *   Allow to choose filename when downloading code
  *   animateLayers() hides the guide to prevent confusion
- *   Added intructions to get date and mouse/touch/keyboard events
+ *   Added instructions to get date, mouse/touch/keyboard events and query color in a pixel as requested by Jordi Losantos and Xavier Paytubi
  *   Allow gridStep to go down to 1
  *   More web app devices compatibility
  *   Ask for confirmation if loading a file will delete unsaved changes